Refactor data source view search tag
