+++ front/lib/swr/data_source_views.ts
@@ -2,9 +2,12 @@ import type {
   ContentNodesViewType,
   DataSourceViewType,
   LightWorkspaceType,
+  TagSearchParams,
 } from "@dust-tt/types";
+import { MIN_SEARCH_QUERY_SIZE } from "@dust-tt/types";
 import { useMemo } from "react";
 import type { Fetcher, KeyedMutator, SWRConfiguration } from "swr";
+import useSWR from "swr";
 
 import type { CursorPaginationParams } from "@app/lib/api/pagination";
 import {
@@ -14,6 +17,10 @@ import {
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
 import type { GetDataSourceViewsResponseBody } from "@app/pages/api/w/[wId]/data_source_views";
+import type {
+  PostTagSearchBody,
+  PostTagSearchResponseBody,
+} from "@app/pages/api/w/[wId]/data_source_views/tags/search";
 import type { GetDataSourceViewContentNodes } from "@app/pages/api/w/[wId]/spaces/[spaceId]/data_source_views/[dsvId]/content-nodes";
 import type { GetDataSourceConfigurationResponseBody } from "@app/pages/api/w/[wId]/spaces/[spaceId]/data_sources/[dsId]/configuration";
 
@@ -203,5 +210,54 @@ export function useDataSourceViewConnectorConfiguration({
     mutateConfiguration: mutate,
     isConfigurationLoading: !disabled && !error && !data,
     isConfigurationError: error,
+  };
+}
+
+export function useDataSourceViewSearchTags({
+  dataSourceViews,
+  disabled = false,
+  owner,
+  query,
+}: {
+  dataSourceViews: DataSourceViewType[];
+  disabled?: boolean;
+  owner: LightWorkspaceType;
+  query: string;
+}) {
+  const url =
+    query.length >= MIN_SEARCH_QUERY_SIZE
+      ? `/api/w/${owner.sId}/data_source_views/tags/search`
+      : null;
+
+  const body: PostTagSearchBody = {
+    query,
+    queryType: "match",
+    dataSourceViewIds: dataSourceViews.map((dsv) => dsv.sId),
+  };
+
+  const fetchKey = JSON.stringify([url, body]);
+
+  const { data, error, mutate, isValidating, isLoading } = useSWRWithDefaults(
+    fetchKey,
+    async () => {
+      if (!url) {
+        return null;
+      }
+
+      return fetcherWithBody([url, body, "POST"]);
+    },
+    {
+      revalidateOnFocus: false,
+      revalidateOnReconnect: false,
+      disabled,
+    }
+  );
+
+  return {
+    tags: useMemo(() => data?.tags || [], [data]),
+    isLoading,
+    isError: !!error,
+    mutate,
+    isValidating,
   };
 }
